+++ TypographyKit/Classes/PropertyListParsingService.swift
@@ -7,9 +7,9 @@
 //
 
 struct PropertyListParsingService: ParsingService {
+    
     func parse(_ data: Data) -> ParsingServiceResult? {
-        guard let plistDictionary = try? PropertyListSerialization.propertyList(from: data,
-                                                                                options: [],
+        guard let plistDictionary = try? PropertyListSerialization.propertyList(from: data, options: [],
                                                                                 format: nil)
             as? [String: Any],
             let result = plistDictionary else {
@@ -17,4 +17,5 @@ struct PropertyListParsingService: ParsingService {
         }
         return parse(result)
     }
+    
 }